requirements.txt
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
PyMuPDF>=1.23.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
tiktoken>=0.5.0
llama-cpp-python>=0.2.0
huggingface-hub>=0.20.0
